requirements.txt
python-telegram-bot==20.8
yt-dlp>=2023.10.13
python-dotenv
curl_cffi
telethon
cryptg
